rendering: remove the OpenGL requirement
This change technically makes the program use software rendering instead of hardware rendering, thus putting more work on the shoulders of the CPU.

The reasoning behind this change is that rendering using SDL requires a lot of memory to run (160 MB). This change removes this big memory requirement, decreasing memory usage to 17 MB on linux and to a beautiful number of 4.8 MB (after one test) on windows.